Add device sharing infrastructure
Based on group of devices (namespaces / access ctrl tokens),
and implicit groups created by harvesting Omlet.

And lot of infrastructure to make Omlet more prominent in
the engine, while still keeping an abstraction layer in
case we don't actually want Omlet.

+++ engine/db/devices.js
@@ -269,7 +269,7 @@ module.exports = new lang.Class({
         }.bind(this));
 
         this._devices[device.uniqueId] = device;
-        if (addToDB) {
+        if (addToDB && !device.isTransient) {
             var state = device.serialize();
             var uniqueId = device.uniqueId;
             return this._syncdb.insertOne(uniqueId,
@@ -289,9 +289,13 @@ module.exports = new lang.Class({
 
     removeDevice: function(device) {
         delete this._devices[device.uniqueId];
-        return this._syncdb.deleteOne(device.uniqueId).then(function() {
+        if (device.isTransient) {
             this._notifyDeviceRemoved(device);
-        }.bind(this));
+        } else {
+            return this._syncdb.deleteOne(device.uniqueId).then(function() {
+                this._notifyDeviceRemoved(device);
+            }.bind(this));
+        }
     },
 
     hasDevice: function(uniqueId) {
